refactor(accelerator): q7e3-F-07 split versions/mod.rs into focused submodules
versions/mod.rs was a ~1000-LOC module mixing version policy, cache layout,
release metadata, and 626 lines of tests (incl. 9 downloader tests stranded
outside downloader.rs, reachable only via a cfg(test) re-export). Split into
version_policy.rs (NetworkTier, AztecVersion, is_valid_version, eviction),
cache_layout.rs (base dir / bb path / cache scan), release_metadata.rs
(platform naming, download URLs, http_client, sha256, the SEC-02-caveated
GitHub digest lookup); mod.rs is now an 18-line re-export hub so every
external versions::X path is unchanged (the F-12 lesson). downloader.rs
imports its deps explicitly per submodule instead of the 9-item 'use super'
and absorbs its 9 stranded tests (the cfg(test) re-export is deleted); the
stranded download_bb doc comment left behind by the earlier extraction is
dropped. Tests move with their code. core exactly 133/133 (no test lost);
clippy --all-targets -D warnings clean.
